perf(text): isolate baker diagnostics
