+++ docs/log.md
@@ -2,6 +2,7 @@
 
 ## 2026-08-01
 
+- **Thin baker diagnostics boundary** — Moved direct Wasm timing and memory observation behind a private diagnostic-only TypeScript entry while retaining the Rust phase observer behind its non-default `profiling` feature. The production package-size build now rejects diagnostic module or symbol reachability, clock calls in thin baker hosts, and profiling/timing Wasm imports or exports; packed consumers receive no diagnostic module, and the package concept records exact thin-build, diagnostic-run, and evidence-refresh commands.
 - **MTSDF phase attribution** — Added a package-owned small, medium, complete, and combined profiler over the shared optimized native, direct-Wasm, and serial-Worker bake paths. Authenticated measurements isolate texel generation as the dominant phase while separately recording selection, outlines, packing, KTX2, GLB, transfer, and memory high-water evidence. The profiling-only TypeScript and Rust entry points stay outside production execution; the rebased Darwin arm64 production MTSDF baker measures 552,025 raw / 215,027 gzip / 169,041 Brotli bytes.
 - **Bounded-coverage size remediation** — Removed the second derived Serde serialization graph from Bitmap and MTSDF coverage-capable bakers while preserving strict seed validation and byte-identical canonical descriptors. The measured Darwin arm64 Wasm payloads now occupy 626,940 raw / 234,735 gzip / 180,503 Brotli bytes for Bitmap and 553,190 raw / 215,142 gzip / 169,365 Brotli bytes for MTSDF. Dedicated pre-coverage growth gates bound the remaining strict decoder, cmap-resolution, and canonical-policy cost. Bitmap runtime decode now also derives the canonical policy key from authenticated strikes and coverage before creating GPU resources, matching the existing MTSDF boundary.
