+++ packages/text/tests/package/esm-only.test.mjs
@@ -13,6 +13,7 @@ test('the published contract is ESM-only', async () => {
   assert.equal(manifest.main, undefined);
   assert.equal(manifest.module, undefined);
   assert.deepEqual(manifest.bin, { 'pmndrs-text-bake': './dist/node/cli.js' });
+  assert.equal(manifest.exports['./internal/raster-baker-profile'], undefined);
   assert.deepEqual(manifest.pmndrs, {
     text: { bitmap: './bakers/bitmap', msdf: './bakers/msdf', slug: './bakers/slug' },
   });
